+++ base-pi/py_server.py
@@ -4,6 +4,7 @@ import metrics
 import asyncio
 import signal
 from metrics import asyncsshloop, register_metric_events, cpuloop, send_fake_antenna_stats
+from shutdown import register_shutdown_commands
 import sys
 
 
@@ -67,6 +68,7 @@ cpu_started = False
 
 
 register_metric_events(sio)
+register_shutdown_commands(sio)
 # =================== Start Server ===================
 
 @sio.event
