+++ server/py_server.py
@@ -13,8 +13,7 @@ from drive_uart import read_drive_uart_loop, send_drive_heartbeat, register_driv
 from arm import read_arm_can_loop, request_arm_position_loop, register_arm_events
 from camera_pt import register_camera_pt_events
 from gps import ZEDF9P, GPS_Data, GNRMC, read_gps_data, send_fake_gps_data
-
-
+from shutdown import register_shutdown_commands
 # Toggle drive communication transport for testing / fallback
 # True = UART drive path
 # False = original CAN drive path
@@ -321,6 +320,7 @@ register_metric_events(sio)
 register_drive_events(sio,serial_ports)
 register_arm_events(sio, serial_ports)
 register_camera_pt_events(sio,serial_ports)
+register_shutdown_commands(sio)
 
 # =================== Start Server ===================
